Remove ' ' as a legal character.
Technically, it is allowed in dataset names, but it's pretty clearly
going to cause too many problems with stuff like checking whether a
dataset is currently mounted, since zfs list won't say that it is if it
was mounted via mount -t zfs, and the mount command's output isn't set
up in a way to make it easy (or maybe even possible) to correctly parse
in the face of paths with spaces. And even if I got it to mostly work,
there's a high chance that there would be corner cases that wouldn't.
So, it's just easier to disallow it. bectl and beadm seem to disallow
' ' as well.