Fix it so that we're actually splitting on tabs and not all whitespace.
zfs list -H separates by tabs, and yet we've been splitting based on
general whitespace, which is wrong even if it happens to work in most
cases. We're of course still screwed in someone manages to create a
dataset or snapshot with a tab in it, and we're probably still screwed
with other whitespace, but this at least reduces the problem.