Rewrite destroy to improve its performance.
Apparently, zfs list has a clones property, which gives a comma-separate
list of clones for a snapshot. Using that is _much_ faster than using
zfs list on the entire pool and searching for the origin property on
every dataset to see which snapshots have clones, which is what the code
was doing before.