Add the -k flag to destroy.
It will tell bemgr to keep the origin rather than destroy it.

Personally, I hate it when origins are kept around, or the program asks
about it, but it occurs to me that someone may actually want to keep
around the snapshot for some reason, so I'm adding the -k flag to give
that option. However, the default remains to destroy the origin.