Recursively promote to deal with openzfs bug.
https://github.com/openzfs/zfs/issues/15587

In at least the situation where a snapshot of a clone is cloned,
zfs promote gets into a situation where it can take multiple attempts
for the dataset to no longer be a clone. Snapshots get shuffled around
but not sufficiently. This of course causes problems with destroy and
means that activate fails to actually end up with the active BE not
being a clone.

So, we're now going to repeatedly promote the clone until it's no longer
a clone. For now, the cut-off is 100 attempts, which presumably is
ludicrously high, but since I have no clue what the exact logic is on
the part of zfs promote which causes this problem, I have no clue how
many attempts it could take - or if it's even guaranteed to succeed
eventually. So, 100 should be plenty in any sane situation while still
stopping eventually if it's not working.

Also, the snapshot names now include milliseconds. This likely isn't
necessary in practice, but the integration tests are running into issues
where we manage to have two snapshots with the same name on different
datasets due to multiple create calls taking less than a second, and
that causes problems with zfs promote. And we either need to add logic
to detect such a conflict and handle it or make it less likely that the
names won't conflict. The latter is much easier, and this situation
shouldn't happen in practice. It is of course possible that someone
could create it by manually creating snapshots with the same names, but
if so, they can also manually fix it. We can rethink the solution later
if need be, but this should cover all normal situations.