Fix bemgr destroy -F bename@snapshot.
Apparently, zfs destroy ignores -f with snapshots, and of course, zfs
mount does the same thing on on Linux, so we can't fully fix the
situation on Linux, but we can at least try to unmount the snapshot
first (and we can force that on FreeBSD). So, FreeBSD should be fully
fixed in this regard, and Linux is as fixed as we can make it given the
limitations with forcibly unmounting on Linux.