Make it so that -f is passed through on Linux with umount.
This changes it so that bemgr umount -f is allowed on Linux instead of
being an error. It still won't work, because zfs unmount -f won't work,
but it's passed along to zfs umount, so if zfs unmount ever starts
actually supporting -f on linux instead of ignoring it, then it will
just work with bemgr instead of needing updates - though given what the
regular umount command looks like on Linux, I'm skeptical that it will
ever be fixed.