Fix it so that phobos is statically linked.
We don't want to need phobos to be on the system (especially since we'll
then need the correct version). So, this changes it so that we're
linking it statically. Unfortunately, for some reason, with ldc, that
results in having to link in libz, and I can't get it to link statically
even though libz.a is on my system. So, we're still stuck with another
link dependency, but it's one that will probably be on the system
anyway. It's doubly annoying given that I'm pretty sure that the only
reason than libz is needed by Phobos is for std.zip, and bemgr isn't
using that, but I've already wasted too much time trying to fix it, so
I'm leaving it as-is for now.