Fix it so that the tarball install script can handle symlinks.
Arch uses symlinks with some of its man page directories, and
unfortunately, cp -R wants to replace the symlinks with directories even
though all I care about is that the files get copied to the correct
place, and I don't see a way to tell cp to let symlinks stay symlinks.
So, unfortunately, the install script now copies the files explicitly
(and uses mkdir -p for the directories, which does not have the problem
that cp does) instead of being able to have the install script not care
about what the files are. So, if it ever changes, the script will need
to be updated. It probably won't ever change though.