Fix #1: bemgr create fails with time zones that aren't west of UTC.
This makes it so that the automatic snapshots that bemgr creates don't
have a time zone offset in them. It works fine when the offset is
negative, but when it's positive (or zero with a +), the + causes the
snapshot creation to fail, since + is not a legal character in snapshot
names.

It would be nice to retain the offset, but unlike with a program that's
creating regular snapshots (e.g. zfs-auto-snapshot), the snapshots that
bemgr is making don't really need to worry about DST switches or other
time zone changes (really, they just need unique names). So, it just
seemed simpler to remove the offset than to come up with a replacement
for + that was reasonably clear.