Add gendocs.d and related files for generating html documentation.
gendocs.d is a stripped down and tweaked version of the program that I
use to generate documentation for my libraries (e.g.
https://github.com/jmdavis/dxml). For a library, that involves using
ddoc to generate the documentation from the source code, but in this
case, it's using mandoc to generate the documentation from the man page.
However, it still has the ddoc plumbing around it, primarily so that it
can be hooked into the generation of the html for my website,
jmdavisprog.com, and that includes adding a header and footer to the
documentation by providing an additional ddoc file with the appropriate
macros.

Realistically, I expect that no one else will care about this, but
gendocs.d can be run to get the man page as html locally.